+++ test/test_partition2.c
@@ -78,22 +78,22 @@ refine_fn (p4est_t * p4est, p4est_topidx_t which_tree,
   return 1;
 }
 
-static int
+static double
 weight_one (p4est_t * p4est, p4est_topidx_t which_tree,
             p4est_quadrant_t * quadrant)
 {
-  return 1;
+  return 1.0;
 }
 
-static int
+static double
 weight_once (p4est_t * p4est, p4est_topidx_t which_tree,
              p4est_quadrant_t * quadrant)
 {
   if (weight_counter++ == weight_index) {
-    return 1;
+    return 1.0;
   }
 
-  return 0;
+  return 0.0;
 }
 
 static int
